Split widget: adaptive wrap
